moved components into singular components file
