+++ src/components/category_option/category_option.css
