+++ src/components/top_bar/top_bar.js
@@ -1,7 +1,7 @@
 // src/pages/admin/components/top_bar.js
 import React from 'react'
 import { signOut } from 'firebase/auth'
-import { auth } from '../../../firebase_config.js'
+import { auth } from '../../firebase_config.js'
 import { useNavigate } from 'react-router-dom'
 import './top_bar.css'
